Move imported literal export policy into semantics
